--- server/src/api/v1/openai/organization/index.ts
@@ -1,10 +1,0 @@
-import { Express } from "express";
-import Constants from "../../../../constants";
-
-const URL_PREFIX = Constants.OPENAI_ORGANIZATION_PREFIX;
-
-export function registerOrganization(app: Express): void {
-    app.get(`${URL_PREFIX}`, (req, res) => {
-        res.send(URL_PREFIX);
-    });
-}
