+++ server/src/constants.ts
@@ -2,5 +2,5 @@ export default class Constants {
     static API_PREFIX = '/api';
     static V1_PREFIX = `${this.API_PREFIX}/v1`;
     static OPENAI_PREFIX = `${this.V1_PREFIX}/openai`;
-    static OPENAI_ORGANIZATION_PREFIX = `${this.OPENAI_PREFIX}/organization`;
+    static OPENAI_MODELS_PREFIX = `${this.OPENAI_PREFIX}/models`;
 }
