+++ server/src/index.ts
@@ -18,12 +18,3 @@ registerApi(app);
 app.listen(port, () => {
     console.log(`server started at http://localhost:${port}`);
 });
-
-/*
-    TODO: Proxy certain import openAI commands and work on helper library:
-    - https://platform.openai.com/docs/api-reference/requesting-organization
-    Example:
-    curl https://api.openai.com/v1/models \
-      -H "Authorization: Bearer $OPENAI_API_KEY" \
-      -H "OpenAI-Organization: <org-id>"
-*/
